Replace old wizard by the new version
