+++ intellij/src/nl/tudelft/watchdog/intellij/ui/wizards/IdInputPanel.java
@@ -1,4 +1,4 @@
-package nl.tudelft.watchdog.intellij.ui.new_wizards;
+package nl.tudelft.watchdog.intellij.ui.wizards;
 
 import com.intellij.ui.DocumentAdapter;
 import nl.tudelft.watchdog.core.logic.network.NetworkUtils;
@@ -11,8 +11,8 @@ import javax.swing.event.DocumentEvent;
 import java.awt.*;
 import java.util.function.Consumer;
 
-import static nl.tudelft.watchdog.intellij.ui.new_wizards.UserRegistrationStep.ID_LENGTH;
-import static nl.tudelft.watchdog.intellij.ui.new_wizards.WizardStep.DEFAULT_SPACING;
+import static nl.tudelft.watchdog.intellij.ui.wizards.UserRegistrationStep.ID_LENGTH;
+import static nl.tudelft.watchdog.intellij.ui.wizards.WizardStep.DEFAULT_SPACING;
 
 abstract class IdInputPanel extends JPanel {
