+++ intellij/src/nl/tudelft/watchdog/intellij/ui/wizards/UserRegistrationInputPanel.java
@@ -1,4 +1,4 @@
-package nl.tudelft.watchdog.intellij.ui.new_wizards;
+package nl.tudelft.watchdog.intellij.ui.wizards;
 
 import com.intellij.openapi.ui.ComboBox;
 import nl.tudelft.watchdog.core.logic.network.JsonTransferer;
@@ -10,8 +10,8 @@ import javax.swing.*;
 import java.awt.*;
 import java.util.function.Consumer;
 
-import static nl.tudelft.watchdog.intellij.ui.new_wizards.UserRegistrationStep.ID_LENGTH;
-import static nl.tudelft.watchdog.intellij.ui.new_wizards.WizardStep.DEFAULT_SPACING;
+import static nl.tudelft.watchdog.intellij.ui.wizards.UserRegistrationStep.ID_LENGTH;
+import static nl.tudelft.watchdog.intellij.ui.wizards.WizardStep.DEFAULT_SPACING;
 
 class UserRegistrationInputPanel extends JPanel {
