+++ intellij/src/nl/tudelft/watchdog/intellij/ui/wizards/UserWelcomeScreen.java
@@ -1,4 +1,4 @@
-package nl.tudelft.watchdog.intellij.ui.new_wizards;
+package nl.tudelft.watchdog.intellij.ui.wizards;
 
 import javax.swing.*;
 import java.awt.*;
